feat(@eurora/desktop): separate the hotkey change to a separate component
