+++ apps/desktop/src/routes/(onboarding)/onboarding/final/+page.svelte
@@ -5,6 +5,6 @@
 <div class="w-full h-full p-6 flex flex-col justify-between">
 	<h1 class="text-2xl font-bold">Eurora is ready!</h1>
 	<div class="flex justify-start">
-		<Button variant="outline" href="/">Back</Button>
+		<Button variant="outline" href="/onboarding">Back</Button>
 	</div>
 </div>
