+++ apps/desktop/src/routes/settings/hotkey/+page.svelte
@@ -1,4 +1,11 @@
 <script lang="ts">
+	import Hotkey from '$lib/components/Hotkey.svelte';
 </script>
 
-Hotkey
+<div class="w-full h-full p-6 flex flex-col justify-between">
+	<h1 class="text-2xl font-bold">Set up Eurora hotkey</h1>
+
+	<div class="flex w-full h-full items-center justify-center">
+		<Hotkey />
+	</div>
+</div>
